Organize landing page components into dedicated folders
