+++ src/app/pages/landing/components/dropzone/landing-dropzone.component.html
@@ -1,0 +1,23 @@
+<div
+  class="landing__dropzone"
+  role="button"
+  tabindex="0"
+  (click)="onOpenFilePicker()"
+  (keydown)="fileUploadKeydown.emit($event)"
+  (dragover)="fileDragOver.emit($event)"
+  (dragleave)="fileDragLeave.emit($event)"
+  (drop)="fileDrop.emit($event)"
+  [class.landing__dropzone--active]="fileDropActive"
+  [attr.aria-label]="'app.upload.title' | t"
+>
+  <input #pdfFileInput type="file" accept="application/pdf" (change)="fileSelected.emit($event)" hidden />
+  <div class="landing__icon" aria-hidden="true">📄</div>
+  <div class="landing__text">
+    <span class="landing__headline">{{ 'app.upload.title' | t }}</span>
+    <span class="landing__subheadline">{{ 'app.upload.subtitle' | t }}</span>
+  </div>
+  <button type="button" class="landing__cta" (click)="onOpenFilePicker($event)">
+    {{ 'app.landing.cta' | t }}
+  </button>
+  <span class="landing__hint">{{ 'app.upload.hint' | t }}</span>
+</div>
