+++ src/app/pages/landing/components/footer/landing-footer.component.html
@@ -1,0 +1,13 @@
+<footer class="footer landing__footer">
+  <div class="footer__brand">
+    <img class="footer__logo" src="/logo.svg" [attr.alt]="'Logotipo de ' + appName" />
+    <div class="footer__brand-text">
+      <span class="footer__app-name">{{ appName }}</span>
+      <span class="footer__version">v{{ version }}</span>
+    </div>
+  </div>
+  <div class="footer__details">
+    <span>© {{ currentYear }} {{ appName }}</span>
+    <span>Desarrollado por {{ appAuthor }}</span>
+  </div>
+</footer>
