+++ src/app/pages/landing/components/footer/landing-footer.component.ts
@@ -1,0 +1,15 @@
+import { CommonModule } from '@angular/common';
+import { Component, Input } from '@angular/core';
+
+@Component({
+  selector: 'app-landing-footer',
+  standalone: true,
+  imports: [CommonModule],
+  templateUrl: './landing-footer.component.html',
+})
+export class LandingFooterComponent {
+  @Input({ required: true }) appName = '';
+  @Input({ required: true }) version = '';
+  @Input({ required: true }) currentYear!: number;
+  @Input({ required: true }) appAuthor = '';
+}
