+++ src/app/pages/landing/components/hero/landing-hero.component.html
@@ -1,0 +1,5 @@
+<div class="landing__hero">
+  <span class="landing__badge">{{ 'app.landing.badge' | t }}</span>
+  <h1 class="landing__title">{{ 'app.landing.title' | t }}</h1>
+  <p class="landing__description">{{ 'app.landing.description' | t }}</p>
+</div>
