+++ src/app/pages/landing/components/language-bar/landing-language-bar.component.html
@@ -1,0 +1,8 @@
+<div class="landing__language">
+  <app-language-selector
+    [languages]="languages"
+    [selectedLanguage]="selectedLanguage"
+    (languageChange)="languageChange.emit($event)"
+    variant="compact"
+  ></app-language-selector>
+</div>
